server: use res.sendFile instead of fs.createReadStream
This is better in particular to actually handle range requests
correctly.  Before this commit, we actually sent everything to the
client, so advertising partial request support in
64ed4378f0519a2333e569127a57842794d7635c was actually kinda a lie.
With this commit, we _actually_ automatically support range requests,
only sending the part of the file that is requested.
In addition, we get (weak) etags for free!